[requested-changes] Improved docs, refactored code and fixed test
