+++ README.rst
@@ -96,7 +96,8 @@ Available Features
 * CSV Export of monitoring data
 * An overview of the status of the network is shown in the admin dashboard,
   a chart shows the percentages of devices which are online, offline or having issues;
-  timeseries charts are shown for WiFi sessions and Traffic of all managed devices,
+  there are also `two timeseries charts which show the total unique WiFI clients and
+  the traffic flowing to the network <dashboard-monitoring-charts>`_,
   a geographic map is also available for those who use the geographic features of OpenWISP
 * Possibility to configure additional `Metrics <#openwisp_monitoring_metrics>`_ and `Charts <#openwisp_monitoring_charts>`_
 * Extensible active check system: it's possible to write additional checks that
@@ -801,6 +802,22 @@ Mobile Access Technology in use
 .. figure:: https://github.com/openwisp/openwisp-monitoring/raw/docs/docs/access-technology.png
   :align: center
 
+Dashboard Monitoring Charts
+---------------------------
+
+.. figure:: https://github.com/openwisp/openwisp-monitoring/raw/docs/docs/dashboard-charts.png
+  :align: center
+
+OpenWISP Monitoring adds two timeseries charts to the admin dashboard:
+
+- **General WiFi clients Chart**: Shows the number of connected clients to the WiFi
+  interfaces of devices in the network.
+- **General traffic Chart**: Shows the amount of traffic flowing in the network.
+
+You can configure the interfaces included in the **General traffic chart** using
+the `"OPENWISP_MONITORING_DASHBOARD_TRAFFIC_CHART"
+<#openwisp_monitoring_dashboard_traffic_chart>`_ setting.
+
 Monitoring WiFi Sessions
 ------------------------
 
@@ -1233,7 +1250,7 @@ does not have list of interfaces defined in ``OPENWISP_MONITORING_DASHBOARD_TRAF
 
 **Note**: If a user can manage more that one organizations (e.g. superusers),
 then the **General Traffic** chart will always show data from interfaces
-of ``__all__`` configuration. 
+of ``__all__`` configuration.
 
 ``OPENWISP_MONITORING_METRICS``
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -1746,12 +1763,26 @@ and in the `Browsable web page <#browsable-web-interface>`_ of each point,
 here we'll provide just a list of the available endpoints,
 for further information please open the URL of the endpoint in your browser.
 
+Retrieve general monitoring charts
+##################################
+
+.. code-block:: text
+
+    GET /api/v1/monitoring/dashboard/
+
+This API endpoint is used to show dashboard monitoring charts. It supports
+multi-tenancy and allows filtering monitoring data with `organization_slug`, e.g.:
+
+.. code-block:: text
+
+    GET /api/v1/monitoring/dashboard/?organization_slug=<org1-slug>,<org2-slug>
+
 Retrieve device charts and device status data
 #############################################
 
 .. code-block:: text
 
-    GET /v1/monitoring/device/{pk}/?key={key}&status=true
+    GET /api/v1/monitoring/device/{pk}/?key={key}&status=true
 
 The format used for Device Status is inspired by `NetJSON DeviceMonitoring <http://netjson.org/docs/what.html#devicemonitoring>`_.
 
@@ -1763,13 +1794,13 @@ Collect device metrics and status
 
 .. code-block:: text
 
-    POST /v1/monitoring/device/{pk}/?key={key}&time={time}
+    POST /api/v1/monitoring/device/{pk}/?key={key}&time={time}
 
 If data is latest then an additional parameter current can also be passed. For e.g.:
 
 .. code-block:: text
 
-    POST /v1/monitoring/device/{pk}/?key={key}&time={time}&current=true
+    POST /api/v1/monitoring/device/{pk}/?key={key}&time={time}&current=true
 
 The format used for Device Status is inspired by `NetJSON DeviceMonitoring <http://netjson.org/docs/what.html#devicemonitoring>`_.
