+++ openwisp_monitoring/device/api/views.py
@@ -1,21 +1,14 @@
-import csv
 import logging
 import uuid
-from collections import OrderedDict
 from copy import deepcopy
 from datetime import datetime, timedelta
-from io import StringIO
 
-from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.db.models import Count, Q
-from django.http import HttpResponse
 from django.utils.timezone import now
 from django.utils.translation import gettext_lazy as _
 from pytz import UTC
-from pytz import timezone as tz
-from pytz.exceptions import UnknownTimeZoneError
 from rest_framework import serializers, status
 from rest_framework.generics import GenericAPIView
 from rest_framework.response import Response
@@ -29,7 +22,7 @@ from openwisp_controller.geo.api.views import (
 
 from ... import settings as monitoring_settings
 from ...monitoring.configuration import ACCESS_TECHNOLOGIES
-from ...monitoring.exceptions import InvalidChartConfigException
+from ...views import MonitoringApiViewMixin
 from ..schema import schema
 from ..signals import device_metrics_received
 from .serializers import MonitoringDeviceSerializer, MonitoringGeoJsonLocationSerializer
@@ -44,7 +37,7 @@ DeviceData = load_model('device_monitoring', 'DeviceData')
 Location = load_model('geo', 'Location')
 
 
-class DeviceMetricView(GenericAPIView):
+class DeviceMetricView(MonitoringApiViewMixin, GenericAPIView):
     model = DeviceData
     queryset = DeviceData.objects.select_related('devicelocation').all()
     serializer_class = serializers.Serializer
@@ -58,112 +51,18 @@ class DeviceMetricView(GenericAPIView):
         except ValueError:
             return Response({'detail': 'not found'}, status=404)
         self.instance = self.get_object()
+        return super().get(request, pk)
+
+    def _get_charts(self, request, *args, **kwargs):
         ct = ContentType.objects.get_for_model(Device)
-        charts = Chart.objects.filter(
-            metric__object_id=pk, metric__content_type=ct
+        return Chart.objects.filter(
+            metric__object_id=args[0], metric__content_type=ct
         ).select_related('metric')
-        # determine time range
-        time = request.query_params.get('time', Chart.DEFAULT_TIME)
-        if time not in Chart.GROUP_MAP.keys():
-            return Response({'detail': 'Time range not supported'}, status=400)
-        # try to read timezone
-        timezone = request.query_params.get('timezone', settings.TIME_ZONE)
-        try:
-            tz(timezone)
-        except UnknownTimeZoneError:
-            return Response('Unkown Time Zone', status=400)
-        # prepare response data
-        data = self._get_charts_data(charts, time, timezone)
-        # csv export has a different response
-        if request.query_params.get('csv'):
-            response = HttpResponse(self._get_csv(data), content_type='text/csv')
-            response['Content-Disposition'] = 'attachment; filename=data.csv'
-            return response
-        # add device data if requested
+
+    def _get_additional_data(self, request, *args, **kwargs):
         if request.query_params.get('status', False):
-            data['data'] = self.instance.data
-        return Response(data)
-
-    def _get_charts_data(self, charts, time, timezone):
-        chart_map = {}
-        x_axys = True
-        data = OrderedDict({'charts': []})
-        for chart in charts:
-            # prepare chart dict
-            try:
-                chart_dict = chart.read(time=time, x_axys=x_axys, timezone=timezone)
-                if not chart_dict['traces']:
-                    continue
-                chart_dict['description'] = chart.description
-                chart_dict['title'] = chart.title.format(
-                    metric=chart.metric, **chart.metric.tags
-                )
-                chart_dict['type'] = chart.type
-                chart_dict['unit'] = chart.unit
-                chart_dict['summary_labels'] = chart.summary_labels
-                chart_dict['colors'] = chart.colors
-                chart_dict['colorscale'] = chart.colorscale
-                for attr in ['fill', 'xaxis', 'yaxis']:
-                    value = getattr(chart, attr)
-                    if value:
-                        chart_dict[attr] = value
-                if chart.trace_type:
-                    chart_dict['trace_type'] = chart.trace_type
-                if chart.trace_order:
-                    chart_dict['trace_order'] = chart.trace_order
-            except InvalidChartConfigException:
-                logger.exception(f'Skipped chart for metric {chart.metric}')
-                continue
-            # get x axys (only once)
-            if x_axys and chart_dict['x'] and chart.type != 'histogram':
-                data['x'] = chart_dict.pop('x')
-                x_axys = False
-            # prepare to sort the items according to
-            # the order in the chart configuration
-            key = f'{chart.order} {chart_dict["title"]}'
-            chart_map[key] = chart_dict
-        # add sorted chart list to chart data
-        data['charts'] = list(OrderedDict(sorted(chart_map.items())).values())
-        return data
-
-    def _get_csv(self, data):
-        header = ['time']
-        columns = [data.get('x')]
-        histograms = []
-        for chart in data['charts']:
-            if chart['type'] == 'histogram':
-                histograms.append(chart)
-                continue
-            for trace in chart['traces']:
-                header.append(self._get_csv_header(chart, trace))
-                columns.append(trace[1])
-        rows = [header]
-        for index, element in enumerate(data.get('x', [])):
-            row = []
-            for column in columns:
-                row.append(column[index])
-            rows.append(row)
-        for chart in histograms:
-            rows.append([])
-            rows.append([chart['title']])
-            # Export value as 0 if it is None
-            for key, value in chart['summary'].items():
-                if chart['summary'][key] is None:
-                    chart['summary'][key] = 0
-            # Sort Histogram on the basis of value in the descending order
-            sorted_charts = sorted(
-                chart['summary'].items(), key=lambda x: x[1], reverse=True
-            )
-            for field, value in sorted_charts:
-                rows.append([field, value])
-        # write CSV to in-memory file object
-        fileobj = StringIO()
-        csv.writer(fileobj).writerows(rows)
-        return fileobj.getvalue()
-
-    def _get_csv_header(self, chart, trace):
-        header = trace[0]
-        return f'{header} - {chart["title"]}'
+            return {'data': self.instance.data}
+        return {}
 
     def post(self, request, pk):
         self.instance = self.get_object()
