+++ openwisp_monitoring/monitoring/utils.py
@@ -1,0 +1,6 @@
+from django.utils.text import slugify
+
+
+def clean_timeseries_data_key(value):
+    value = value.replace('.', '_')
+    return slugify(value).replace('-', '_')
